code/requirements.txt
fastapi==0.95.1
fastapi_utils==0.2.1
uvicorn[standard]==0.21.1
pymongo==4.3.3
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
loguru==0.7.0
email-validator==1.3.1
redis==4.5.4
python-decouple==3.8
python-multipart==0.0.6
pika==1.3.1
requests==2.28.2
pydantic==1.10.7
pendulum==2.1.2
cachetools==5.3.3

# dev
pytest==7.3.1
httpx==0.24.0
mongomock==4.1.2
pylint-quotes==0.2.3
pylint-pydantic==0.1.8
pytest-randomly==3.12.0